+++ src/app/Components/ComponentsSamples/TabView/TabViewSamples.xaml.cs
@@ -51,4 +51,9 @@ public partial class TabViewSamples
             OnPropertyChanged();
         }
     }
+
+    private void TabViewSample_OnOnSelectedTabIndexChanged(object? sender, DIPS.Mobile.UI.Components.TabView.TabView.TabViewEventArgs e)
+    {
+        
+    }
 }